fastapi_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pillow>=9.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
numpy==1.23.5
httpx>=0.24.1
pydantic-settings>=2.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
deepface>=0.0.79
tensorflow-cpu==2.12.0
keras==2.12.0
mtcnn>=0.1.1
retina-face>=0.0.13
PyJWT>=2.0.0
starlette>=0.27.0 